observe_ai/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.8.5
chromadb==1.3.4
sentence-transformers==5.1.2
langchain==1.0.7
langchain-chroma==1.0.0
langchain-community==0.3.31
langchain-deepseek==1.0.1
langchain-huggingface==1.0.1
langgraph==1.0.3
joblib==1.5.2
neo4j==6.0.3
numpy==2.3.4
pandas==2.3.3
pydantic==2.12.4
tqdm==4.67.1
fastapi==0.123.9
xgboost==3.1.2